// src/types/partials/RequestGroupAction.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TRelatedArtifact } from '../partials/RelatedArtifact';
import { TRequestGroupCondition } from '../partials/RequestGroupCondition';
import { TRequestGroupRelatedAction } from '../partials/RequestGroupRelatedAction';
import { TDateTime } from '../simpleTypes/DateTime';
import { TQuantity } from '../partials/Quantity';
import { TPeriod } from '../partials/Period';
import { TRange } from '../partials/Range';
import { TTiming } from '../partials/Timing';
import { TReference } from '../partials/Reference';

export type TRequestGroupAction = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    prefix?: String;
    title?: String;
    description?: String;
    textEquivalent?: String;
    priority?: String;
    code?: TCodeableConcept[];
    documentation?: TRelatedArtifact[];
    condition?: TRequestGroupCondition[];
    relatedAction?: TRequestGroupRelatedAction[];
    timingDateTime?: TDateTime;
    timingAge?: TQuantity;
    timingPeriod?: TPeriod;
    timingDuration?: TQuantity;
    timingRange?: TRange;
    timingTiming?: TTiming;
    participant?: TReference[];
    type?: TCodeableConcept;
    groupingBehavior?: String;
    selectionBehavior?: String;
    requiredBehavior?: String;
    precheckBehavior?: String;
    cardinalityBehavior?: String;
    resource?: TReference;
    action?: TRequestGroupAction[];
};